+++ src/main/java/tw/kid7/BannerMaker/util/SenderUtil.java
@@ -1,0 +1,15 @@
+package tw.kid7.BannerMaker.util;
+
+import org.bukkit.command.CommandSender;
+import org.bukkit.command.ConsoleCommandSender;
+import org.bukkit.entity.Player;
+
+public class SenderUtil {
+    public static boolean isPlayer(CommandSender sender) {
+        return (sender instanceof Player);
+    }
+
+    public static boolean isConsole(CommandSender sender) {
+        return (sender instanceof ConsoleCommandSender);
+    }
+}